Lithium, Crafting, Sounds, Boxes Added Lithium with its variants. Lithium reacts with water, larger reactions if the item is a block. Added conversion crafting recipes for most of the resources. Added future storage item to store hazardous materials.

+++ src/main/java/net/intercraft/intercraftcore/element/Element.java
@@ -27,7 +27,7 @@ public class Element
      * Element information
      */
     public final String name,     symbol;
-    public final short    tintPrim, tintSec;
+    public final int    tintPrim, tintSec;
     public final ElementComposition[] composition;
 
     /**
@@ -45,8 +45,8 @@ public class Element
     {
         this.name        = name;
         this.symbol      = dictionary.getSymbol();
-        this.tintPrim    = (short) dictionary.getColorPrimary();
-        this.tintSec     = (short) dictionary.getColorSecondary();
+        this.tintPrim    = dictionary.getColorPrimary();
+        this.tintSec     = dictionary.getColorSecondary();
         this.composition = dictionary.getComposition();
     }
